libmount: move mount(2) error messages to legacy hook
