+++ libmount/src/mountP.h
@@ -681,6 +681,8 @@ extern int mnt_context_get_mountinfo(struct libmnt_context *cxt, struct libmnt_t
 extern int mnt_context_get_mountinfo_for_target(struct libmnt_context *cxt,
 				    struct libmnt_table **mountinfo, const char *tgt);
 
+extern struct libmnt_fs *mnt_context_get_already_mounted(struct libmnt_context *cxt);
+
 extern int mnt_context_prepare_srcpath(struct libmnt_context *cxt);
 extern int mnt_context_guess_srcpath_fstype(struct libmnt_context *cxt, char **type);
 extern int mnt_context_guess_fstype(struct libmnt_context *cxt);
